fixed a bug

+++ app.rb
@@ -9,6 +9,11 @@ end
 post '/' do
   doc = params['content']
   name = params['name'].empty?? doc.hash.to_s : params['name']
+
+  while File.exists? "clips/#{name}"  do
+    name = name.hash.to_s
+  end
+
   return haml :index if name.match '/'
 
   begin
@@ -24,7 +29,7 @@ get '/:id' do
   path = "clips/#{params['id']}"
   if File.exists?(path)
     @buf = File.open(path) { |f| f.read }
-    File.delete(path)
+    File.delete(path) if not request.env['HTTP_USER_AGENT'].match 'facebook'
     haml :paste
   else
     haml :nopaste
